requirements.txt
anthropic>=0.30.0
playwright>=1.44.0
flask>=3.0.0
requests>=2.32.0
python-dotenv>=1.0.0
pillow>=10.0.0
